1.0.2

+++ application/models/presupuestos_model.php
@@ -45,14 +45,15 @@
 		}
 	}
 	
-	function deleteArticulo($id_articulo=0){
-		if($id_articulo!=0){
-			$query = $this->db->query("SELECT *	FROM presupuesto_detalle WHERE id_articulo='$id_articulo'");
+	function deleteArticulo($id_articulo=0, $id_presupuesto=0){
+		if($id_articulo!=0 && $id_presupuesto!=0){
+			$query = $this->db->query("SELECT *	FROM presupuesto_detalle WHERE id_articulo='$id_articulo' AND id_presupuesto='$id_presupuesto'");
 			if($query->num_rows()>0){
 				$data= array(            
 					'id_articulo'=>$id_articulo,
 				);
 				$this->db->where('id_articulo', $id_articulo);
+				$this->db->where('id_presupuesto', $id_presupuesto);
 				$this->db->delete('presupuesto_detalle');
 
 				if($this->db->affected_rows()>0){
@@ -68,35 +69,38 @@
 		}
 	}
 	
-	function getArticulos($id_presupuesto=0){
-		
-		$query = $this->db->query("SELECT 
-	    articulo.id_articulo as id,
-		articulo.articulo,
-		articulo.descripcion,
-		articulo.precio,
-		articulo.stock,
-		articulo.stock_min,
-		articulo.stock_max,
-		articulo.imagen,
-		articulo.id_categoria,
-		presupuesto_detalle.id_articulo, 
-		presupuesto_detalle.id_presupuesto, 
-		presupuesto_detalle.cantidad
-		FROM presupuesto_detalle 
-		INNER JOIN articulo 
-		ON(presupuesto_detalle.id_articulo=articulo.id_articulo) 
-		WHERE id_presupuesto='$id_presupuesto' 
-		ORDER BY articulo.articulo");
-		
-		if($query->num_rows() > 0){
-			foreach ($query->result() as $fila){
-				$data[] = $fila;
-			}
-			return $data;
+	function getArticulos($id_presupuestos){
+		if(is_array($id_presupuestos)){
+			//crear un nuevo presupuesto
 		}else{
-			return FALSE;
-		}
+			$query = $this->db->query("SELECT 
+		    articulo.id_articulo as id,
+			articulo.articulo,
+			articulo.descripcion,
+			articulo.precio,
+			articulo.stock,
+			articulo.stock_min,
+			articulo.stock_max,
+			articulo.imagen,
+			articulo.id_categoria,
+			presupuesto_detalle.id_articulo, 
+			presupuesto_detalle.id_presupuesto, 
+			presupuesto_detalle.cantidad
+			FROM presupuesto_detalle 
+			INNER JOIN articulo 
+			ON(presupuesto_detalle.id_articulo=articulo.id_articulo) 
+			WHERE id_presupuesto='$id_presupuestos' 
+			ORDER BY articulo.articulo");
+			
+			if($query->num_rows() > 0){
+				foreach ($query->result() as $fila){
+					$data[] = $fila;
+				}
+				return $data;
+			}else{
+				return FALSE;
+			}	
+		}				
 	}
 
 /**************************************************************************
